eoc drops more stars qb drops more herb bags eow/boc drop more ore plantera drops more chlorophyte ml drops more luminite fixed bbb tooltip, all summon weapons display emode nerf did something to hopefully fix dps display issue with big sting nerfed cobalt ench, now does 20/40/100 damage and scales with ranged damage instead of using the damage of whatever spawned it nerfed devi light balls, they persist for half a second longer and accelerate marginally slower HOPEFULLY fixed eri timestop obliterating servers in mp eri timestop invaders now have startup accel removed eridanus DR eri drops loot per player in mp (confirm every player can get different enches from each other) nerfed tiki ench, summon weapons now attack slower when tiki effect is applying fixed tiki taking your number of minion projectiles, not the actual number of slots used by minions fixed siblings glitching out and spawning ~30 copies of their visual displays constantly fixed mutant true eye not firing deathray fixed mutant lighting issues in mp martian probes and certain pillar enemies are immune to suffocation penetrator dive lasts even longer, should be enough to dive from top of large world to bottom added a toggle for supersonic black belt buffed heart of eternal, gives 10% damage/crit (was 5%) champions besides eri have slight life scaling in mp standardized sell price on slime slinging slasher, diffractor blaster nerfed rockeater launcher nerfed big sting buffed regurgitator buffed mech leash mech leash flail and eyes have trails now phanleash flail has glowing trail and bolts die earlier changed implementation of hover in flight mastery soul and upgrades to hopefully prevent mp desync

+++ Projectiles/DeviBoss/DeviBoss.cs
@@ -37,13 +37,15 @@ namespace FargowiltasSouls.Projectiles.DeviBoss
                 projectile.Center = Main.npc[ai1].Center;
                 projectile.alpha = Main.npc[ai1].alpha;
                 projectile.direction = projectile.spriteDirection = Main.npc[ai1].direction;
-                projectile.timeLeft = 2;
+                projectile.timeLeft = 30;
 
-                projectile.frame = (int)(Main.npc[ai1].frame.Y / (float)(Main.projectileTexture[projectile.type].Height / Main.projFrames[projectile.type]));
+                if (!Main.dedServ)
+                    projectile.frame = (int)(Main.npc[ai1].frame.Y / (float)(Main.projectileTexture[projectile.type].Height / Main.projFrames[projectile.type]));
             }
             else
             {
-                projectile.Kill();
+                if (Main.netMode != NetmodeID.MultiplayerClient)
+                    projectile.Kill();
                 return;
             }
         }
